Storing State in Cookies
Added a test to ensure that the `balance` state of the app will be set
from a cookie by default. In essence, this test checks the reducer
that sets the balance to ensure that it will set a proper balance for
default cases where there is not action being dispatched.